Add small code example how to add base-app-url
I tried to follow this guide but I found it a bit unclear how exactly to add `base-app-url` in the manifest. The text says explains it a bit, but is `<admin>` nested into `<meta>`? The linked article "manifest" doesn't include the admin part and I also confused it with `<registrationUrl>`.

So I added a small code snippet which will hopefully help to understand it more easily.